# src/main/resources/erp-api-contract/erp-openapi-contract.yaml
openapi: 3.0.1
info:
  title: Enterprise Resource Planning Manager
  description: API contract for the Enterprise Resource Planning Manager
  contact:
    name: DevOps Factory
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - BEARER_JWT: [ ]
tags:
  - name: user
    description: User resource.
  - name: country
    description: Country resource.
  - name: customer
    description: potential customer
paths:
  /users:
    post:
      tags:
        - user
      description: Create a new user.
      operationId: createUser
      responses:
        "201":
          $ref: '#/components/responses/UserDetailsResponse'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
  /users/current-user:
    get:
      tags:
        - user
      description: Retrieve the current authenticated user’s details.
      operationId: getCurrentUser
      responses:
        "200":
          $ref: '#/components/responses/UserDetailsResponse'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
  /users/current-user/roles:
    get:
      tags:
        - user
      description: Retrieve the roles assigned to the current authenticated user.
      operationId: getCurrentUserRoles
      responses:
        "200":
          $ref: '#/components/responses/ListRolesResponse'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
  /countries:
    get:
      tags:
        - country
      description: Get all the countries available.
      operationId: getCountriesAvailable
      responses:
        "200":
          $ref: './responses/country/CountriesResponse.yaml'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
  /customers:
    get:
      tags:
        - customer
      summary: Search customers
      operationId: searchCustomers
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: './schemas/customer/CustomerStatus.yaml'
        - name: ownerId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: customerId
          example: customerId,asc

      responses:
        "200":
          $ref: './responses/customer/CustomersPageResponse.yaml'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
    post:
      tags:
        - customer
      summary: Create customer
      description: Create a new customer.
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/customer/CustomerModel.yaml'
      responses:
        "201":
          $ref: './responses/customer/CustomerResponse.yaml'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
  /customers/{id}:
    get:
      tags:
        - customer
      summary: Get customer by ID
      description: Retrieve a customer by its ID.
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          description: Customer ID.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: './responses/customer/CustomerResponse.yaml'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
    put:
      tags:
        - customer
      summary: Update customer
      description: Update an existing customer.
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          description: Customer ID.
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: './requests/customer/UpdateCustomerRequest.yaml'
      responses:
        "200":
          $ref: './responses/customer/CustomerResponse.yaml'
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
    delete:
      tags:
        - customer
      summary: Delete customer
      description: Delete a customer by its ID.
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          required: true
          description: Customer ID.
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Customer successfully deleted.
        "400":
          $ref: './responses/common/BadRequest.yaml'
        "401":
          $ref: './responses/common/Unauthorized.yaml'
        "403":
          $ref: './responses/common/Forbidden.yaml'
        "404":
          $ref: './responses/common/NotFound.yaml'
        "409":
          $ref: './responses/common/Conflict.yaml'
        "500":
          $ref: './responses/common/InternalServerError.yaml'
components:
  schemas:
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserModel'
    UserModel:
      type: object
      properties:
        businessKey:
          type: integer
          format: int64
        username:
          type: string
        givenName:
          type: string
        familyName:
          type: string
        email:
          type: string
        preferredLanguage:
          type: string
          example: en-US
      description: User model.

    RolesResponse:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
      description: List of roles.
  responses:
    UserDetailsResponse:
      description: Successfully retrieved user details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserResponse'
    ListRolesResponse:
      description: Successfully retrieved list of roles.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RolesResponse'
    BadRequest:
      description: The request was invalid or cannot be served. The exact error should be explained in the error payload.
      content:
        application/json:
          schema:
            $ref: './schemas/common/ProblemDetail.yaml'
  securitySchemes:
    BEARER_JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT